Use fakeroot state file to preserve device attribute
In my previous patch for support for NFS boot using fakeroot, i had forgot to make it save the fakeroot environment state.
Because of this it was losing the device attribute on files, and they were not bootable.
This patch fix the issue.

Also put the fakeroot state file in git ignore list.